# Remove global rand lock in skiplist