--- a/skl/skl.go
+++ b/skl/skl.go
@@ -36,6 +36,7 @@ import (
 	"math"
 	"math/rand"
 	"sync/atomic"
+	"time"
 	"unsafe"
 
 	"github.com/dgraph-io/badger/v2/y"
@@ -79,6 +80,7 @@ type Skiplist struct {
 	head   *node
 	ref    int32
 	arena  *Arena
+	rng    rand.Source
 }
 
 // IncrRef increases the refcount
@@ -132,6 +134,7 @@ func NewSkiplist(arenaSize int64) *Skiplist {
 		head:   head,
 		arena:  arena,
 		ref:    1,
+		rng:    rand.NewSource(time.Now().UnixNano()),
 	}
 }
 
@@ -165,9 +168,9 @@ func (s *node) casNextOffset(h int, old, val uint32) bool {
 //	return n != nil && y.CompareKeys(key, n.key) > 0
 //}
 
-func randomHeight() int {
+func (s *Skiplist) randomHeight() int {
 	h := 1
-	for h < maxHeight && rand.Uint32() <= heightIncrease {
+	for h < maxHeight && uint32(s.rng.Int63()) <= heightIncrease {
 		h++
 	}
 	return h
@@ -300,7 +303,7 @@ func (s *Skiplist) Put(key []byte, v y.ValueStruct) {
 	}
 
 	// We do need to create a new node.
-	height := randomHeight()
+	height := s.randomHeight()
 	x := newNode(s.arena, key, v, height)
 
 	// Try to increase s.height via CAS.