The math/rand package (https://golang.org/src/math/rand/rand.go) uses a global lock to allow concurrent access to the rand object. We do not write to skiplist concurrently and hence we don't need a lock at all.

This blog has some interesting details https://blog.sgmansfield.com/2016/01/the-hidden-dangers-of-default-rand/

The PR replaces `math.Rand` with `ristretto/z.FastRand()`. `FastRand` is much faster than `math.Rand` and `rand.New(..)` generators.

The change improves concurrent writes to skiplist by ~30%
```go
func BenchmarkWrite(b *testing.B) {
	value := newValue(123)
	l := NewSkiplist(int64((b.N + 1) * MaxNodeSize))
	defer l.DecrRef()
	b.ResetTimer()
	b.RunParallel(func(pb *testing.PB) {
		rng := rand.New(rand.NewSource(time.Now().UnixNano()))
		for pb.Next() {
			l.Put(randomKey(rng), y.ValueStruct{Value: value, Meta: 0, UserMeta: 0})
		}
	})
}
```
```
name      old time/op  new time/op  delta
Write-16   657ns ± 3%   441ns ± 1%  -32.94%  (p=0.000 n=9+10)
```


This PR also fixes the `test timed out after 10 minutes` issue on teamcity (see build https://teamcity.dgraph.io/viewLog.html?buildId=36889&buildTypeId=Badger_UnitTests#testNameId521499096021883325 )
The problem here is that the test `TestValueGCManaged` tries to create many bufferes with random data
https://github.com/dgraph-io/badger/blob/ea01d380cedea306b4c9e0e04d072c60cd55e973/value_test.go#L119-L124

The `rand.Read(...)` acquires a global lock (https://golang.org/src/math/rand/rand.go) and since `N=10,000` in the above test, it leads to starvation. We would have too many `rand.Read(...)` calls waiting to acquire the same lock. At the same time, every insert to skip list also needs to acquire the same lock 
https://github.com/dgraph-io/badger/blob/ea01d380cedea306b4c9e0e04d072c60cd55e973/skl/skl.go#L168-L174

Hence, the test wouldn't complete in 10 minutes and the build fails.
<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/badger/1173)
<!-- Reviewable:end -->