@balajijinnah 2019-12-25
we write concurrently while updating head i guess