@jarifibrahim 2019-12-25
> we write concurrently while updating head i guess

We don't concurrently update the same skip list. When we update head, the memtable being used for new writes is a different one. See this

https://github.com/dgraph-io/badger/blob/ea01d380cedea306b4c9e0e04d072c60cd55e973/db.go#L869-L884

Note how we created a new memtable after pushing the existing one to the flush chan.