@coocood 2020-01-10
Why not use `runtime.fastrand`?