@jarifibrahim 2020-01-10
> Why not use `runtime.fastrand`?

Interesting. I did some benchmarks and `fastRand` seems to be more than 2X faster than `rand.NewSource(..)`
```
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/ristretto/z
BenchmarkFastRand-16      	1000000000	         0.292 ns/op
BenchmarkRandSource-16    	1000000000	         0.747 ns/op
BenchmarkRandGlobal-16    	 6822332	       176 ns/op
BenchmarkRandAtomic-16    	77950322	        15.4 ns/op
PASS
ok  	github.com/dgraph-io/ristretto/z	4.808s
```
https://github.com/dgraph-io/ristretto/pull/118/files#diff-a9ed67d3dcec1ff808f43c15fe731caeR289-R310

I'll change it. Thanks @coocood !